user-admin in progress

+++ app/Http/ApiControllers/AdminController.php
@@ -135,7 +135,7 @@ class AdminController extends BaseController
        ];
        $result = $this->adminInterface->store($admin);
        if (isset($result)){
-          event(new ContentCRUDEvent('Create Admin', $request->admin_id, 'Promote', 'Gave admin privileges to '. $result->id));
+          event(new ContentCRUDEvent('Create Admin', $request->admin_id, 'Promote', 'Gave admin privileges to '. $result->name));
           $result = new AdminResource($result);
           $this->data(array('user' => $result));
           return $this->response('201');
@@ -169,15 +169,31 @@ class AdminController extends BaseController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+      // dd($request->all());
+      $validator = Validator::make($request->all(), [
+            'admin_id'  =>  'required|exists:admin,id'
+        ]);
+
+       if ($validator->fails()) {
+           $this->setError('400');
+           $messages=[];
+
+           foreach ($validator->messages()->toArray() as $key=>$value) {
+                $messages[] = (object)['attribue' => $key, 'message' => $value[0]];
+           }
+
+           $this->setValidationError(['validation' => $messages]);
+           return $this->response('400');
+       }
         $admin = $this->adminInterface->find($id);
         if (empty($admin)) {
-           event(new ContentCRUDEvent('Remove Admin', $request->admin_id, 'Promote', 'Remove admin privileges from '. $id));
             $this->setError('404', $id);
             return $this->response('404');
         }else{
             $this->adminInterface->destroy($id);
+            event(new ContentCRUDEvent('Remove Admin', $request->admin_id, 'Demote', 'Remove admin privileges from '. $admin->name));
             $this->data(array('deleted' =>  1));
             return $this->response('200');
         }
